Refactor tests to be scattered across codebase?
