+++ src/risk_free_model.rs
@@ -1,5 +1,7 @@
 use chrono::prelude::Utc;
 use chrono::DateTime;
+#[cfg(test)]
+use chrono::TimeZone;
 
 use log::debug;
 
@@ -27,4 +29,30 @@ impl RiskFreeModel for AnnualisedRiskFreeRate {
         let diff_years: f64 = (diff_secs as f64) / NUMBER_OF_SECONDS_IN_A_YEAR;
         start_value * EULERS_NUMBER.powf(self.ir_exponent * diff_years)
     }
+}
+
+#[test]
+fn one_year_forward_test() {
+    let rfm = get_annualised_risk_free_rate(5.0);
+    let start: f64 = 100.0;
+    let begin_date = Utc.timestamp_millis_opt(1688917143000).unwrap();
+    let end_date = Utc.timestamp_millis_opt(1720539543000).unwrap();
+    let ret = rfm.apply(start, begin_date, end_date);
+    let upper_bound: f64 = 105.02;
+    let lower_bound: f64 = 105.01;
+    assert!(ret < upper_bound);
+    assert!(ret > lower_bound);
+}
+
+#[test]
+fn one_year_backward_test() {
+    let rfm = get_annualised_risk_free_rate(5.0);
+    let start: f64 = 105.0;
+    let begin_date = Utc.timestamp_millis_opt(1720539543000).unwrap();
+    let end_date = Utc.timestamp_millis_opt(1688917143000).unwrap();
+    let ret = rfm.apply(start, begin_date, end_date);
+    let upper_bound: f64 = 99.99;
+    let lower_bound: f64 = 99.98;
+    assert!(ret < upper_bound);
+    assert!(ret > lower_bound);
 }
